rename packages to com.palantir.ls.*
