+++ groovy-language-server/src/main/java/com/palantir/ls/groovy/CompilerWrapper.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.palantir.groovylanguageserver;
+package com.palantir.ls.groovy;
 
 import io.typefox.lsapi.Diagnostic;
 import io.typefox.lsapi.ReferenceParams;
