+++ groovy-language-server/src/main/java/com/palantir/ls/groovy/GroovyLanguageServer.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.palantir.groovylanguageserver;
+package com.palantir.ls.groovy;
 
 import com.google.common.io.Files;
 import io.typefox.lsapi.InitializeParams;
