+++ groovy-language-server/src/main/java/com/palantir/ls/groovy/SingleCompilerWrapperProvider.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.palantir.groovylanguageserver;
+package com.palantir.ls.groovy;
 
 import com.google.common.base.Preconditions;
