+++ groovy-language-server/src/test/java/com/palantir/ls/groovy/GroovycWrapperTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.palantir.groovylanguageserver;
+package com.palantir.ls.groovy;
 
 import static org.hamcrest.Matchers.is;
 import static org.junit.Assert.assertEquals;
@@ -25,8 +25,8 @@ import static org.junit.Assert.assertTrue;
 import com.google.common.base.Optional;
 import com.google.common.collect.Iterables;
 import com.google.common.collect.Sets;
-import com.palantir.groovylanguageserver.util.DefaultDiagnosticBuilder;
-import com.palantir.groovylanguageserver.util.Ranges;
+import com.palantir.ls.util.DefaultDiagnosticBuilder;
+import com.palantir.ls.util.Ranges;
 import io.typefox.lsapi.Diagnostic;
 import io.typefox.lsapi.DiagnosticSeverity;
 import io.typefox.lsapi.Location;
